requirements.txt
pyrogram
TgCrypto
databases[aiosqlite]
aiosqlite[user_db]
python-dotenv
